+++ Content/Blueprints/BP_PlayerController.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3a529da2d0657e541b58e67a368b5954839fb96f92cc4cb6770c6fb3675c8a36
-size 23066
+oid sha256:35be7be6ffc18b1e31c7c2d04e0a5e1cc71262cfecebca9beefe556cb1d472b1
+size 70569
