+++ Source/UE5VoxelTutorial/Private/Voxel/Chunk/MarchingChunk.cpp
@@ -3,7 +3,7 @@
 
 #include "MarchingChunk.h"
 
-#include "FastNoiseLite.h"
+#include "Voxel/Utils/FastNoiseLite.h"
 
 void AMarchingChunk::Setup()
 {
