+++ Source/UE5VoxelTutorial/Private/Voxel/Chunk/NaiveChunk.h
@@ -5,7 +5,8 @@
 #include "CoreMinimal.h"
 
 #include "ChunkBase.h"
-#include "Enums.h"
+#include "Voxel/Utils/Enums.h"
+
 #include "NaiveChunk.generated.h"
 
 class FastNoiseLite;
