+++ Source/UE5VoxelTutorial/UE5VoxelTutorial.Build.cs
@@ -1,5 +1,7 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
+using System.IO;
+
 using UnrealBuildTool;
 
 public class UE5VoxelTutorial : ModuleRules
